refactor: 구분자 분리 및 텍스트 분리
